feat: update ts support
